Move chartattributes to according package
