--- src/main/java/org/openbase/bco/bcozy/controller/powerterminal/Granularity.java
@@ -1,11 +1,0 @@
-package org.openbase.bco.bcozy.controller.powerterminal;
-
-public enum Granularity {
-    CONSUMER, ROOM, OVERALL;
-
-
-    @Override
-    public String toString() {
-        return super.toString().substring(0, 1) + super.toString().substring(1).toLowerCase();
-    }
-}
