+++ src/main/java/org/openbase/bco/bcozy/controller/powerterminal/chartattributes/VisualizationType.java
@@ -1,7 +1,7 @@
-package org.openbase.bco.bcozy.controller.powerterminal;
+package org.openbase.bco.bcozy.controller.powerterminal.chartattributes;
 
-public enum Unit {
-    ENERGY, LIGHTBULB, MONEY;
+public enum VisualizationType {
+    BAR, PIE, WEBVIEW, LINECHART;
 
     @Override
     public String toString() {
